+++ src/transaction/creation_common.rs
@@ -21,7 +21,8 @@ pub(super) enum DisplayOrder {
     ShowJsonArgExamples,
     NodeAddress,
     SecretKey,
-    Input,
+    SpeculativeExec,
+    TransactionPath,
     Output,
     Force,
     Target,
@@ -402,6 +403,7 @@ pub(super) fn apply_common_creation_options(
     subcommand: Command,
     require_secret_key: bool,
     include_node_address: bool,
+    include_transaction_args: bool,
 ) -> Command {
     let mut subcommand = subcommand
         .next_line_help(true)
@@ -432,6 +434,12 @@ pub(super) fn apply_common_creation_options(
             signed, for example by running the `sign-transaction` subcommand.",
         )
     };
+
+    if include_transaction_args {
+        subcommand = subcommand
+            .arg(arg_simple::session::arg())
+            .arg(args_json::session::arg())
+    }
 
     subcommand = subcommand
         .arg(secret_key_arg)
@@ -533,7 +541,7 @@ pub(super) mod transaction_path {
             .short(ARG_SHORT)
             .value_name(ARG_VALUE_NAME)
             .help(ARG_HELP)
-            .display_order(DisplayOrder::Input as usize)
+            .display_order(DisplayOrder::TransactionPath as usize)
     }
 
     pub fn get(matches: &ArgMatches) -> Option<&str> {
@@ -915,9 +923,11 @@ pub(super) mod add_bid {
 
     pub const NAME: &str = "add-bid";
 
+    const ACCEPT_SESSION_ARGS: bool = false;
+
     const ABOUT: &str = "Creates a new add-bid transaction";
     pub fn build() -> Command {
-        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false)
+        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false, ACCEPT_SESSION_ARGS)
     }
 
     pub fn put_transaction_build() -> Command {
@@ -942,7 +952,7 @@ pub(super) mod add_bid {
             amount,
         };
 
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
 
         Ok((params, transaction_str_params))
     }
@@ -961,9 +971,11 @@ pub(super) mod withdraw_bid {
 
     pub const NAME: &str = "withdraw-bid";
 
+    const ACCEPT_SESSION_ARGS: bool = false;
+
     const ABOUT: &str = "Creates a new withdraw-bid transaction";
     pub fn build() -> Command {
-        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false)
+        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false, ACCEPT_SESSION_ARGS)
     }
 
     pub fn put_transaction_build() -> Command {
@@ -980,7 +992,7 @@ pub(super) mod withdraw_bid {
         let amount = transaction_amount::parse_transaction_amount(amount_str)?;
 
         let params = TransactionBuilderParams::WithdrawBid { public_key, amount };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
 
         Ok((params, transaction_str_params))
     }
@@ -997,10 +1009,12 @@ pub(super) mod delegate {
 
     pub const NAME: &str = "delegate";
 
+    const ACCEPT_SESSION_ARGS: bool = false;
+
     const ABOUT: &str = "Creates a new delegate transaction";
 
     pub fn build() -> Command {
-        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false)
+        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false, ACCEPT_SESSION_ARGS)
     }
 
     pub fn put_transaction_build() -> Command {
@@ -1024,7 +1038,7 @@ pub(super) mod delegate {
             validator,
             amount,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
 
         Ok((params, transaction_str_params))
     }
@@ -1043,10 +1057,12 @@ pub(super) mod undelegate {
 
     pub const NAME: &str = "undelegate";
 
+    const ACCEPT_SESSION_ARGS: bool = false;
+
     const ABOUT: &str = "Creates a new delegate transaction";
 
     pub fn build() -> Command {
-        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false)
+        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false, ACCEPT_SESSION_ARGS)
     }
 
     pub fn put_transaction_build() -> Command {
@@ -1070,7 +1086,7 @@ pub(super) mod undelegate {
             validator,
             amount,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
         Ok((params, transaction_str_params))
     }
 
@@ -1087,10 +1103,11 @@ pub(super) mod redelegate {
 
     pub const NAME: &str = "redelegate";
 
+    const ACCEPT_SESSION_ARGS: bool = false;
     const ABOUT: &str = "Creates a new delegate transaction";
 
     pub fn build() -> Command {
-        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false)
+        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false, ACCEPT_SESSION_ARGS)
     }
 
     pub fn put_transaction_build() -> Command {
@@ -1118,7 +1135,7 @@ pub(super) mod redelegate {
             new_validator,
             amount,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
         Ok((params, transaction_str_params))
     }
 
@@ -1136,6 +1153,7 @@ pub(super) mod invocable_entity {
     use casper_client::cli::{CliError, TransactionBuilderParams};
 
     pub const NAME: &str = "invocable-entity";
+    const ACCEPT_SESSION_ARGS: bool = true;
 
     const ABOUT: &str = "Creates a new transaction targeting an invocable entity";
 
@@ -1162,7 +1180,7 @@ pub(super) mod invocable_entity {
             entity_addr,
             entry_point,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
         Ok((params, transaction_str_params))
     }
 
@@ -1177,6 +1195,8 @@ pub(super) mod invocable_entity_alias {
     use casper_client::cli::{CliError, TransactionBuilderParams};
 
     pub const NAME: &str = "invocable-entity-alias";
+
+    const ACCEPT_SESSION_ARGS: bool = true;
 
     const ABOUT: &str = "Creates a new transaction targeting an invocable entity via its alias";
 
@@ -1201,7 +1221,7 @@ pub(super) mod invocable_entity_alias {
             entity_alias,
             entry_point,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
         Ok((params, transaction_str_params))
     }
 
@@ -1216,6 +1236,8 @@ pub(super) mod package {
     use casper_client::cli::{CliError, TransactionBuilderParams};
 
     pub const NAME: &str = "package";
+
+    const ACCEPT_SESSION_ARGS: bool = true;
 
     const ABOUT: &str = "Creates a new transaction targeting a package";
 
@@ -1244,7 +1266,7 @@ pub(super) mod package {
             maybe_entity_version,
             entry_point,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
         Ok((params, transaction_str_params))
     }
 
@@ -1260,6 +1282,8 @@ pub(super) mod package_alias {
     use casper_client::cli::{CliError, TransactionBuilderParams};
 
     pub const NAME: &str = "package-alias";
+
+    const ACCEPT_SESSION_ARGS: bool = true;
 
     const ABOUT: &str = "Creates a new transaction targeting package via its alias";
 
@@ -1288,7 +1312,7 @@ pub(super) mod package_alias {
             maybe_entity_version,
             entry_point,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
         Ok((params, transaction_str_params))
     }
 
@@ -1305,6 +1329,8 @@ pub(super) mod session {
     use casper_client::cli::{CliError, TransactionBuilderParams};
 
     pub const NAME: &str = "session";
+
+    const ACCEPT_SESSION_ARGS: bool = true;
 
     const ABOUT: &str = "Creates a new transaction for running session logic";
 
@@ -1340,7 +1366,7 @@ pub(super) mod session {
             transaction_bytes,
             entry_point,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
         Ok((params, transaction_str_params))
     }
 
@@ -1358,10 +1384,12 @@ pub(super) mod transfer {
 
     pub const NAME: &str = "transfer";
 
+    const ACCEPT_SESSION_ARGS: bool = false;
+
     const ABOUT: &str = "Creates a new native transfer transaction";
 
     pub fn build() -> Command {
-        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false)
+        apply_common_creation_options(add_args(Command::new(NAME).about(ABOUT)), false, false, ACCEPT_SESSION_ARGS)
     }
 
     pub fn put_transaction_build() -> Command {
@@ -1395,7 +1423,7 @@ pub(super) mod transfer {
             maybe_to,
             maybe_id,
         };
-        let transaction_str_params = build_transaction_str_params(matches);
+        let transaction_str_params = build_transaction_str_params(matches, ACCEPT_SESSION_ARGS);
 
         Ok((params, transaction_str_params))
     }
@@ -1457,6 +1485,41 @@ pub(super) mod target {
             .get_one::<String>(ARG_NAME)
             .map(String::as_str)
             .unwrap_or_else(|| panic!("should have {} arg", ARG_NAME))
+    }
+}
+
+/// Handles providing the arg for speculative execution.
+pub(super) mod speculative_exec {
+    use super::*;
+
+    const ARG_NAME: &str = "speculative-exec";
+    const ARG_VALUE_NAME: &str = "HEX STRING OR INTEGER";
+    const ARG_HELP: &str =
+        "If the receiving node supports this, execution of the deploy will only be attempted on \
+        that single node. Full validation of the deploy is not performed, and successful execution \
+        at the given global state is no guarantee that the deploy will be able to be successfully \
+        executed if put to the network, nor should execution costs be expected to be identical. \
+        Optionally provide the hex-encoded block hash or height of the block to specify the global \
+        state on which to execute";
+    const DEFAULT_MISSING_VALUE: &str = "";
+
+    pub(in crate::transaction) fn arg() -> Arg {
+        Arg::new(ARG_NAME)
+            .long(ARG_NAME)
+            .required(false)
+            .value_name(ARG_VALUE_NAME)
+            .num_args(0..=1)
+            .default_missing_value(DEFAULT_MISSING_VALUE)
+            .help(ARG_HELP)
+            .display_order(DisplayOrder::SpeculativeExec as usize)
+    }
+
+    // get: The command line posibilities are encoded by a combination of option and &str
+    // None represents no --speculative-exec argument at all
+    // Some("") represents a --speculative-exec with no/empty argument
+    // Some(block_identifier) represents "--speculative-exec block_identifier"
+    pub(in crate::transaction) fn get(matches: &ArgMatches) -> Option<&str> {
+        matches.get_one::<String>(ARG_NAME).map(String::as_str)
     }
 }
 
@@ -1525,10 +1588,10 @@ pub(super) mod destination_account {
 }
 pub(super) fn apply_common_args(subcommand: Command) -> Command {
     let subcommand = apply_common_args_options(subcommand);
-    apply_common_creation_options(subcommand, false, false)
+    apply_common_creation_options(subcommand, false, false, true)
 }
 
-pub(super) fn build_transaction_str_params(matches: &ArgMatches) -> TransactionStrParams {
+pub(super) fn build_transaction_str_params(matches: &ArgMatches, obtain_session_args: bool) -> TransactionStrParams {
     let secret_key = common::secret_key::get(matches).unwrap_or_default();
     let timestamp = timestamp::get(matches);
     let ttl = ttl::get(matches);
@@ -1536,23 +1599,40 @@ pub(super) fn build_transaction_str_params(matches: &ArgMatches) -> TransactionS
     let maybe_pricing_mode = pricing_mode::get(matches);
     let payment_amount = payment_amount::get(matches);
 
-    let session_args_simple = arg_simple::session::get(matches);
-    let session_args_json = args_json::session::get(matches);
-
     let maybe_output_path = output::get(matches).unwrap_or_default();
     let initiator_addr = initiator_address::get(matches);
-    TransactionStrParams {
-        secret_key,
-        timestamp,
-        ttl,
-        chain_name,
-        initiator_addr,
-        session_args_simple,
-        session_args_json,
-        maybe_pricing_mode,
-        output_path: maybe_output_path,
-        payment_amount,
+
+    if obtain_session_args {
+        let session_args_simple = arg_simple::session::get(matches);
+        let session_args_json = args_json::session::get(matches);
+        TransactionStrParams {
+            secret_key,
+            timestamp,
+            ttl,
+            chain_name,
+            initiator_addr,
+            session_args_simple,
+            session_args_json,
+            maybe_pricing_mode,
+            output_path: maybe_output_path,
+            payment_amount,
+        }
+    } else {
+
+        TransactionStrParams {
+            secret_key,
+            timestamp,
+            ttl,
+            chain_name,
+            initiator_addr,
+            maybe_pricing_mode,
+            output_path: maybe_output_path,
+            payment_amount,
+            ..Default::default()
+        }
     }
+
+
 }
 pub(super) fn add_rpc_args(subcommand: Command) -> Command {
     subcommand