+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.  The format
 * Add `sign-transaction` command for signing transactions to support node 2.0.0.
 * Add `put-transaction` command for sending transactions to support node 2.0.0.
 * Add support for new node RPC method `state_get_entity`, used in the binary's new `get-entity` subcommand.
+* Add `send-transaciton` command for sending previously made and signed transaction files to the network.
 
 ### Changed
 * Update to match change to node RPC `info_get_deploy`.